agent-llm/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.8.2
requests==2.32.3
python-dotenv==1.0.1
streamlit==1.36.0
#These versions are stable with Python 3.12. If you hit environment conflicts, you can relax pins (e.g., fastapi>=0.111,<0.112).